Add PhotoPicker ActivityResultContract
Apply spotless
