+++ filesystem/src/main/AndroidManifest.xml
@@ -15,7 +15,6 @@
   ~ limitations under the License.
   -->
 
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.google.modernstorage.filesystem">
+<manifest package="com.google.modernstorage.filesystem">
 
 </manifest>
