+++ storage/src/androidTest/AndroidManifest.xml
@@ -1,5 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     package="com.google.modernstorage.storage.test">
 
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
