+++ storage/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.google.modernstorage.storage">
+<manifest package="com.google.modernstorage.storage">
 
 </manifest>
